Tests/LibWeb/Text/input/idl-handling-of-null-attribute.html

<script src="include.js"></script>
<script id="script-element">
    test(() => {
        let s = document.getElementById("script-element");
        println(`initial type = '${s.type}'`);
        s.type = null;
        println(`after setting to null = '${s.type}'`);
    });
</script>
