Tests/LibWeb/Text/expected/idl-handling-of-null-attribute.txt

initial type = ''
after setting to null = 'null'
